HParam UI: Move types from scalar_card_types to data_table/types
