+++ tensorboard/webapp/metrics/store/metrics_reducers.ts
@@ -50,7 +50,7 @@ import {
   ColumnHeader,
   ColumnHeaderType,
   DataTableMode,
-} from '../views/card_renderer/scalar_card_types';
+} from '../../widgets/data_table/types';
 import {
   buildOrReturnStateWithPinnedCopy,
   buildOrReturnStateWithUnresolvedImportedPins,
