+++ tensorboard/webapp/metrics/store/metrics_selectors_test.ts
@@ -26,7 +26,7 @@ import {
   createTimeSeriesData,
 } from '../testing';
 import {HistogramMode, TooltipSort, XAxisType} from '../types';
-import {DataTableMode} from '../views/card_renderer/scalar_card_types';
+import {DataTableMode} from '../../widgets/data_table/types';
 import * as selectors from './metrics_selectors';
 import {CardFeatureOverride, MetricsState} from './metrics_types';
