+++ tensorboard/webapp/metrics/views/card_renderer/scalar_card_container.ts
@@ -98,8 +98,6 @@ import {CardRenderer} from '../metrics_view_types';
 import {getTagDisplayName} from '../utils';
 import {DataDownloadDialogContainer} from './data_download_dialog_container';
 import {
-  ColumnHeader,
-  DataTableMode,
   MinMaxStep,
   PartialSeries,
   PartitionedSeries,
@@ -107,8 +105,12 @@ import {
   ScalarCardPoint,
   ScalarCardSeriesMetadataMap,
   SeriesType,
-  SortingInfo,
 } from './scalar_card_types';
+import {
+  ColumnHeader,
+  DataTableMode,
+  SortingInfo,
+} from '../../../widgets/data_table/types';
 import {
   maybeClipTimeSelectionView,
   partitionSeries,
