+++ tensorboard/webapp/metrics/views/right_pane/scalar_column_editor/BUILD
@@ -31,8 +31,8 @@ tf_ng_module(
         "//tensorboard/webapp/metrics:types",
         "//tensorboard/webapp/metrics/actions",
         "//tensorboard/webapp/metrics/store",
-        "//tensorboard/webapp/metrics/views/card_renderer:scalar_card_types",
         "//tensorboard/webapp/widgets/data_table:data_table_header",
+        "//tensorboard/webapp/widgets/data_table:types",
         "@npm//@angular/common",
         "@npm//@angular/core",
         "@npm//@ngrx/store",
@@ -55,8 +55,8 @@ tf_ts_library(
         "//tensorboard/webapp/angular:expect_ngrx_store_testing",
         "//tensorboard/webapp/metrics/actions",
         "//tensorboard/webapp/metrics/store",
-        "//tensorboard/webapp/metrics/views/card_renderer:scalar_card_types",
         "//tensorboard/webapp/widgets/data_table:data_table_header",
+        "//tensorboard/webapp/widgets/data_table:types",
         "@npm//@angular/core",
         "@npm//@angular/platform-browser",
         "@npm//@ngrx/store",
