+++ tensorboard/webapp/metrics/views/right_pane/scalar_column_editor/scalar_column_editor_container.ts
@@ -27,7 +27,7 @@ import {
   getTableEditorSelectedTab,
 } from '../../../store/metrics_selectors';
 import {HeaderEditInfo, HeaderToggleInfo} from '../../../types';
-import {DataTableMode} from '../../card_renderer/scalar_card_types';
+import {DataTableMode} from '../../../../widgets/data_table/types';
 
 @Component({
   selector: 'metrics-scalar-column-editor',
