+++ tensorboard/webapp/persistent_settings/_data_source/BUILD
@@ -24,7 +24,7 @@ tf_ng_module(
     ],
     deps = [
         "//tensorboard/webapp/metrics:types",
-        "//tensorboard/webapp/metrics/views/card_renderer:scalar_card_types",
+        "//tensorboard/webapp/widgets/data_table:types",
     ],
 )
 
@@ -54,7 +54,7 @@ tf_ng_module(
         ":types",
         "//tensorboard/webapp/angular:expect_angular_core_testing",
         "//tensorboard/webapp/metrics:types",
-        "//tensorboard/webapp/metrics/views/card_renderer:scalar_card_types",
+        "//tensorboard/webapp/widgets/data_table:types",
         "@npm//@angular/core",
         "@npm//@types/jasmine",
         "@npm//rxjs",
